server/requirements.txt
fastapi
uvicorn
psycopg2-binary
neo4j
pydantic
pydantic-settings
groq
networkx
scipy
pandas
openpyxl
python-dotenv
python-multipart
transformers
torch
langchain
langchain-groq
langchain-community
pillow